Move Pulse related activities/fragments to pulse package
