+++ app/src/main/java/org/gdg/frisbee/android/activity/GdgNavDrawerActivity.java
@@ -52,6 +52,7 @@ import org.gdg.frisbee.android.arrow.ArrowActivity;
 import org.gdg.frisbee.android.eventseries.TaggedEventSeries;
 import org.gdg.frisbee.android.eventseries.TaggedEventSeriesActivity;
 import org.gdg.frisbee.android.gde.GdeActivity;
+import org.gdg.frisbee.android.pulse.PulseActivity;
 import org.gdg.frisbee.android.task.Builder;
 import org.gdg.frisbee.android.task.CommonAsyncTask;
 import org.gdg.frisbee.android.utils.PrefUtils;
